requirements.txt
flask==2.3.3
Werkzeug==2.3.7
pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
gunicorn==21.2.0
gdown==4.7.1
